move invalid dates to hook
